Add additional Rubies using Workflows.
A recent issue (#1010) highlighted an deprecation under certain Ruby
versions that we should be aware of. So far, we've been using CircleCI
for running tests and previously they didn't support multiple builds.

With 2.0, Workflows allows for something like this to be setup. This
uses YAML inheritance to share a job definition across multiple build
types (defined as different Docker configurations).

This introduces Ruby 2.3 in addition to 2.4.1.